Only attempt to load config after the user has logged in
